[Events] Brought Event inline with the changes to interfaces\Event and made it abstract.
It now employs a simplistic (but flexible) approach to event naming in an inheritance structure, using a public class constant as optional override. This also allows for easy checking of the event type when registering listeners for it, without having to instantiate the class or reflect on it.

However, in situations where the event type is actually dynamic (for example, includes an identifier of an enclosed entity) the class constant becomes superficial and the interface should be implemented directly instead.

Signed-off-by: Michał Chojnacki <m.chojnacki@muyo.io>